Replace `clang-check` with `clang-tidy` (#3131)
Also enforce rules implemented in #3123, #3125, #3126.

+++ src/melee/gr/inlines.h
@@ -153,7 +153,7 @@ Ground_UpdateStarFoxSequence(Ground_GObj* gobj, Ground* gp,
         if (grCorneria_801E2598(gp->u.starfox.xC4.word,
                                 gp->u.starfox.arwing_slot))
         {
-            if ((s32) gp->u.starfox.xCC-- < 0) {
+            if ((gp->u.starfox.xCC--) < 0) {
                 Ground_GObj* sequence_gobj = create_gobj(sequence_gobj_id);
                 grCorneria_801E2738(
                     sequence_gobj, &GET_GROUND(sequence_gobj)->u,
@@ -161,7 +161,7 @@ Ground_UpdateStarFoxSequence(Ground_GObj* gobj, Ground* gp,
                 gp->u.starfox.arwing_slot++;
                 gp->u.starfox.xCC = 0;
             }
-        } else if ((s32) gp->u.starfox.xCC-- < 0) {
+        } else if ((gp->u.starfox.xCC--) < 0) {
             Ground_801C4A08(gobj);
         }
     }
